refactor some systems
