+++ packages/engine/src/xr/XRControllerFunctions.ts
@@ -100,41 +100,3 @@ export const cleanXRInputs = (entity) => {
     }
   })
 }
-
-export const updateXRControllerAnimations = (inputSource: XRInputSourceComponentType) => {
-  const world = Engine.instance.currentWorld
-  const mixers = [inputSource.controllerGripLeft.userData.mixer, inputSource.controllerGripRight.userData.mixer]
-
-  for (const mixer of mixers) {
-    if (!mixer) continue
-    mixer.update(world.deltaSeconds)
-  }
-}
-
-const playTriggerAction = (action, timeScale) => {
-  const time = action.time
-  action.reset()
-  action.time = time
-  action.timeScale = timeScale
-  action.play()
-}
-
-/**
- * Make a fist on controller's trigger button press
- * @param controller XR controller grip
- */
-export const playTriggerPressAnimation = (controller: Group): void => {
-  if (!controller.userData.actions) return
-  const fistAction = controller.userData.actions[0]
-  playTriggerAction(fistAction, 1)
-}
-
-/**
- * Go back to normal pose on controllers trigger button release
- * @param controller XR controller grip
- */
-export const playTriggerReleaseAnimation = (controller: Group) => {
-  if (!controller.userData.actions) return
-  const fistAction = controller.userData.actions[0]
-  playTriggerAction(fistAction, -1)
-}
