+++ packages/engine/src/xr/XRControllerSystem.ts
@@ -12,11 +12,16 @@ import {
 
 import { createActionQueue, getState } from '@xrengine/hyperflux'
 
+import { BinaryValue } from '../common/enums/BinaryValue'
+import { LifecycleValue } from '../common/enums/LifecycleValue'
 import { Engine } from '../ecs/classes/Engine'
 import { Entity } from '../ecs/classes/Entity'
 import { World } from '../ecs/classes/World'
 import { defineQuery, getComponent, hasComponent, setComponent } from '../ecs/functions/ComponentFunctions'
 import { createEntity, removeEntity } from '../ecs/functions/EntityFunctions'
+import { GamepadAxis } from '../input/enums/InputEnums'
+import { InputType } from '../input/enums/InputType'
+import { GamepadMapping } from '../input/functions/GamepadInput'
 import { setVelocityComponent, VelocityComponent } from '../physics/components/VelocityComponent'
 import { EngineRenderer } from '../renderer/WebGLRendererSystem'
 import { addObjectToGroup } from '../scene/components/GroupComponent'
@@ -134,6 +139,49 @@ const updateInputSource = (entity: Entity, space: XRSpace, referenceSpace: XRRef
   if (pose.angularVelocity) velocity.angular.copy(pose.angularVelocity)
 }
 
+export function updateGamepadInput(source: XRInputSource) {
+  if (source.gamepad?.mapping === 'xr-standard') {
+    const mapping = GamepadMapping['xr-standard'][source.handedness]
+
+    source.gamepad.buttons.forEach((button, index) => {
+      // TODO : support button.touched and button.value
+      const prev = Engine.instance.currentWorld.prevInputState.has(mapping[index])
+      if (!prev && !button.pressed) return
+      Engine.instance.currentWorld.inputState.set(mapping[index], {
+        type: InputType.BUTTON,
+        value: [button.pressed ? BinaryValue.ON : BinaryValue.OFF],
+        lifecycleState: button.pressed ? LifecycleValue.Started : LifecycleValue.Ended
+      })
+    })
+
+    // TODO: we shouldn't be modifying input data here, deadzone should be handled elsewhere
+    const inputData = [...source.gamepad.axes]
+    for (let i = 0; i < inputData.length; i++) {
+      if (Math.abs(inputData[i]) < 0.05) inputData[i] = 0
+    }
+
+    // NOTE: we are inverting input here, as the avatar model is flipped 180 degrees. when that is solved, uninvert these gamepad inputs
+    if (inputData.length >= 2) {
+      const Touchpad = source.handedness === 'left' ? GamepadAxis.LTouchpad : GamepadAxis.RTouchpad
+
+      Engine.instance.currentWorld.inputState.set(Touchpad, {
+        type: InputType.TWODIM,
+        value: [inputData[0], inputData[1]],
+        lifecycleState: LifecycleValue.Started // TODO
+      })
+    }
+
+    if (inputData.length >= 4) {
+      const Thumbstick = source.handedness === 'left' ? GamepadAxis.LThumbstick : GamepadAxis.RThumbstick
+      Engine.instance.currentWorld.inputState.set(Thumbstick, {
+        type: InputType.TWODIM,
+        value: [inputData[2], inputData[3]],
+        lifecycleState: LifecycleValue.Started // TODO
+      })
+    }
+  }
+}
+
 export const xrInputSourcesMap = new Map<XRInputSource, Entity>()
 
 export default async function XRControllerSystem(world: World) {
@@ -215,6 +263,9 @@ export default async function XRControllerSystem(world: World) {
       EngineRenderer.instance.xrSession.addEventListener('inputsourceschange', onInputSourcesChange)
 
     if (Engine.instance.xrFrame) {
+      const session = Engine.instance.xrFrame.session
+      for (const source of session.inputSources) updateGamepadInput(source)
+
       /**
        * @todo
        * when the session starts, we cant easily add the inputsourceschange event when we need it, so we lazily check all input sources here
